requirements.txt
# Dependências do Dr. CLT - Advogado Trabalhista Virtual
parlant>=3.0.2
supabase>=2.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0